Put into same file batch use cases
